A useful getLength method is added to ol.proj
A new ol.proj.getLength method returns the length of a lineString in
metres for EPSG:4326 and EPSG:3857. If Projections with near constant
scale such as EPSG:27700 or UTM are constructed, they can be tagged as
'constantScale'. If this is done then ol.proj.getLength will perform
simple cartesian calculations rather than using the Haversine formula.
The 'constantScale' attribute has obviated the need for customising the
getPointResolution function. ol.proj.getPointResolution is modified to
return an unchanged resolution for 'constantSacle' projections.
'constantScale' is set true for projections with units of pixels.